*Set up way to edit Meals with Ingredients
Need to figure out how to autorefresh edit's recyclerview

+++ app/src/main/java/com/kunle/shoppinglistapp/data/MealWithIngredientsDao.java
@@ -27,7 +27,7 @@ public interface MealWithIngredientsDao {
     void updatePair(MealFoodMap crossRef);
 
     @Query("DELETE FROM MealFoodMap WHERE mealId = :mealID")
-    void deleteMealIngredients(Integer mealID);
+    void deleteMealIngredients(Long mealID);
 
     @Query("DELETE FROM MealFoodMap")
     void deleteAllMealsWithIngredients();
